add: access all segmentations in all viewports in compare
