+++ platform/ui-next/src/components/SegmentationTable/SegmentationSegments.tsx
@@ -4,249 +4,7 @@ import { PopupCard, PopupCardTrigger, PopupCardContent } from '../PopupCard';
 import { useSegmentationTableContext, useSegmentationExpanded } from './contexts';
 import { SegmentStatistics } from './SegmentStatistics';
 import { useDynamicMaxHeight } from '../../hooks/useDynamicMaxHeight';
-
-// Simple Dialog component for the modal
-const Dialog = ({ isOpen, onClose, children }) => {
-  if (!isOpen) {
-    return null;
-  }
-  return (
-    <div className="fixed inset-0 z-50 flex items-center justify-center bg-black bg-opacity-50 text-white">
-      <div className="bg-bkg-low max-h-[80vh] w-full max-w-4xl overflow-y-auto rounded-lg p-6">
-        <div className="mb-4 flex items-center justify-between">
-          <h2 className="text-xl font-semibold">Compare Segments Statistics</h2>
-          <button
-            onClick={onClose}
-            className="text-muted-foreground hover:text-white"
-          >
-            ✕
-          </button>
-        </div>
-        {children}
-      </div>
-    </div>
-  );
-};
-
-const CompareSegs = ({ data }) => {
-  const [isModalOpen, setIsModalOpen] = useState(false);
-  const [leftSegment, setLeftSegment] = useState('');
-  const [rightSegment, setRightSegment] = useState('');
-
-  const handleNumber = value => {
-    if (value === null || value === undefined) {
-      return '';
-    }
-    if (Array.isArray(value)) {
-      return value.map(handleNumber).join(', ');
-    }
-    return Number(value).toFixed(2);
-  };
-
-  const allSegments = data.flatMap(entry =>
-    Object.entries(entry.segmentation.segments).map(([key, segment]) => ({
-      key,
-      label: segment.label,
-      segment,
-    }))
-  );
-
-  const leftSelected = allSegments.find(seg => seg.key === leftSegment)?.segment;
-  const rightSelected = allSegments.find(seg => seg.key === rightSegment)?.segment;
-  console.log('leftSelected:', leftSelected);
-  return (
-    <>
-      <div className="m-2 flex items-center justify-center">
-        <button
-          onClick={() => setIsModalOpen(true)}
-          className="rounded bg-blue-600 p-2 text-white hover:bg-blue-700"
-        >
-          Compare Stats
-        </button>
-      </div>
-      <Dialog
-        isOpen={isModalOpen}
-        onClose={() => setIsModalOpen(false)}
-      >
-        <div
-          className="rounded-lg p-6"
-          style={{ backgroundColor: 'rgb(9, 12, 42)' }}
-        >
-          <div className="space-y-6">
-            <div className="mb-4 flex space-x-4">
-              <div className="flex-1">
-                <label
-                  className="mb-1 block text-sm font-medium"
-                  style={{ color: 'rgb(125, 179, 207)' }}
-                >
-                  Left Segment
-                </label>
-                <select
-                  value={leftSegment}
-                  onChange={e => setLeftSegment(e.target.value)}
-                  className="w-full rounded border border-gray-700 bg-gray-800 px-2 py-1 text-white"
-                >
-                  <option value="">Select a segment</option>
-                  {allSegments.map(seg => (
-                    <option
-                      key={seg.key}
-                      value={seg.key}
-                      className="text-black"
-                    >
-                      {seg.label}
-                    </option>
-                  ))}
-                </select>
-              </div>
-              <div className="flex-1">
-                <label
-                  className="mb-1 block text-sm font-medium"
-                  style={{ color: 'rgb(125, 179, 207)' }}
-                >
-                  Right Segment
-                </label>
-                <select
-                  value={rightSegment}
-                  onChange={e => setRightSegment(e.target.value)}
-                  className="w-full rounded border border-gray-700 bg-gray-800 px-2 py-1 text-white"
-                >
-                  <option value="">Select a segment</option>
-                  {allSegments.map(seg => (
-                    <option
-                      key={seg.key}
-                      value={seg.key}
-                      className="text-black"
-                    >
-                      {seg.label}
-                    </option>
-                  ))}
-                </select>
-              </div>
-            </div>
-
-            <div className="grid grid-cols-2 gap-6">
-              <div>
-                {leftSelected ? (
-                  <div
-                    className="border-b pb-4"
-                    style={{ borderColor: 'rgb(37, 42, 116)' }}
-                  >
-                    <div className="mb-3 flex items-center space-x-2">
-                      <h3
-                        className="text-lg font-semibold"
-                        style={{ color: 'rgb(125, 179, 207)' }}
-                      >
-                        {leftSelected.label}
-                      </h3>
-                    </div>
-                    <div className="space-y-2">
-                      {leftSelected?.cachedStats?.namedStats &&
-                        Object.entries(leftSelected.cachedStats.namedStats)
-                          .filter(([_, stat]) => stat && stat.value !== null)
-                          .sort((a, b) => {
-                            const orderA = a[1]?.order ?? Number.MAX_SAFE_INTEGER;
-                            const orderB = b[1]?.order ?? Number.MAX_SAFE_INTEGER;
-                            return orderA - orderB;
-                          })
-                          .map(([statKey, stat]) => (
-                            <div
-                              key={statKey}
-                              className="flex justify-between text-sm"
-                            >
-                              <div
-                                className="text-lg"
-                                style={{ color: 'rgb(125, 179, 207)' }}
-                              >
-                                {stat.label}
-                              </div>
-                              <div className="flex items-baseline space-x-1">
-                                <span
-                                  style={{ color: 'white' }}
-                                  className="text-lg font-normal"
-                                >
-                                  {handleNumber(stat.value)}
-                                </span>
-                                <span
-                                  style={{ color: 'rgb(125, 179, 207)' }}
-                                  className="text-md"
-                                >
-                                  {stat.unit && stat.unit !== 'none' ? stat.unit : ''}
-                                </span>
-                              </div>
-                            </div>
-                          ))}
-                    </div>
-                  </div>
-                ) : (
-                  <p className="text-gray-500">Select a segment to compare</p>
-                )}
-              </div>
-
-              <div>
-                {rightSelected ? (
-                  <div
-                    className="border-b pb-4"
-                    style={{ borderColor: 'rgb(37, 42, 116)' }}
-                  >
-                    <div className="mb-3 flex items-center space-x-2">
-                      <h3
-                        className="text-lg font-semibold"
-                        style={{ color: 'rgb(125, 179, 207)' }}
-                      >
-                        {rightSelected.label}
-                      </h3>
-                    </div>
-                    <div className="space-y-2">
-                      {rightSelected?.cachedStats?.namedStats &&
-                        Object.entries(rightSelected.cachedStats.namedStats)
-                          .filter(([_, stat]) => stat && stat.value !== null)
-                          .sort((a, b) => {
-                            const orderA = a[1]?.order ?? Number.MAX_SAFE_INTEGER;
-                            const orderB = b[1]?.order ?? Number.MAX_SAFE_INTEGER;
-                            return orderA - orderB;
-                          })
-                          .map(([statKey, stat]) => (
-                            <div
-                              key={statKey}
-                              className="flex justify-between text-sm"
-                            >
-                              <div
-                                className="text-lg"
-                                style={{ color: 'rgb(125, 179, 207)' }}
-                              >
-                                {stat.label}
-                              </div>
-                              <div className="flex items-baseline space-x-1 text-lg">
-                                <span
-                                  style={{ color: 'white' }}
-                                  className="text-lg font-normal"
-                                >
-                                  {handleNumber(stat.value)}
-                                </span>
-                                <span
-                                  style={{ color: 'rgb(125, 179, 207)' }}
-                                  className="text-md"
-                                >
-                                  {stat.unit && stat.unit !== 'none' ? stat.unit : ''}
-                                </span>
-                              </div>
-                            </div>
-                          ))}
-                    </div>
-                  </div>
-                ) : (
-                  <p className="text-gray-500">Select a segment to compare</p>
-                )}
-              </div>
-            </div>
-          </div>
-        </div>
-      </Dialog>
-    </>
-  );
-};
-
-export default CompareSegs;
+import CompareSegs from './CompareSegs';
 
 export const SegmentationSegments = ({ children = null }: { children?: React.ReactNode }) => {
   const {
@@ -259,15 +17,31 @@ export const SegmentationSegments = ({ children = null }: { children?: React.Rea
     onSegmentEdit,
     onSegmentDelete,
     data,
+    AllSegmentations,
   } = useSegmentationTableContext('SegmentationSegments');
 
-  data.map((entry, entryIndex) => {
-    console.log(`\nENTRYYYYYYYY DATA ${entryIndex}:`);
-    for (const [key, segment] of Object.entries(entry.segmentation.segments)) {
-      const namedStats = segment?.cachedStats?.namedStats;
-      console.log(`  Segment ${key} namedStats:`, namedStats || 'No namedStats for this segment.');
-    }
+  console.log('AllSegmentations WASAAAAl: ', AllSegmentations);
+  console.log('WHOLE DATA :', data);
+
+  AllSegmentations.forEach((segmentation, index) => {
+    console.log(`${segmentation.label}:`);
+    const segments = segmentation.segments;
+
+    Object.keys(segments).forEach(segmentKey => {
+      const segment = segments[segmentKey];
+      if (segment.cachedStats && segment.cachedStats.namedStats) {
+        console.log(`  Segment ${segmentKey} nameStats:`, segment.cachedStats.namedStats);
+      }
+    });
   });
+
+  // data.map((entry, entryIndex) => {
+  //   console.log(`\nENTRYYYYYYYY DATA ${entryIndex}:`);
+  //   for (const [key, segment] of Object.entries(entry.segmentation.segments)) {
+  //     const namedStats = segment?.cachedStats?.namedStats;
+  //     console.log(`  Segment ${key} namedStats:`, namedStats || 'No namedStats for this segment.');
+  //   }
+  // });
 
   let segmentation;
   let representation;
@@ -286,7 +60,10 @@ export const SegmentationSegments = ({ children = null }: { children?: React.Rea
 
   const segments = Object.values(representation.segments);
   const isActiveSegmentation = segmentation.segmentationId === activeSegmentationId;
-  console.log('segments: ', segments);
+  // console.log('segments: ', segments);
+
+  // const allSegs = SegmentationService?.getSegmentations();
+  // console.log('allSegs: ', allSegs);
 
   const { ref: scrollableContainerRef, maxHeight } = useDynamicMaxHeight(segments);
 
@@ -389,7 +166,7 @@ export const SegmentationSegments = ({ children = null }: { children?: React.Rea
           })}
         </div>
       </ScrollArea>
-      <CompareSegs data={data} />
+      <CompareSegs AllSegmentations={AllSegmentations} />
     </div>
   );
 };
